# Improve performance of `Oj.load` with symbol_keys mode